requirements.txt
neptune_contrib==0.25.0
pandas==1.1.3
opencv_python==4.4.0.46
tensorflow==2.3.1
neptune_client==0.4.129
